+++ BlitzCache/BlitzCache.cs
@@ -18,7 +18,8 @@ namespace BlitzCacheCore
         /// <param name="cleanupInterval">Interval for automatic cleanup of unused semaphores (default: 10 seconds)</param>
         /// <param name="maxTopSlowest">Max number of top slowest queries to store (0 for improved performance) (default: 5 queries)</param>
         /// <param name="maxTopHeaviest">Max number of heaviest entries to track (0 disables). Default: 5.</param>
-        public BlitzCache(long? defaultMilliseconds = 60000, TimeSpan? cleanupInterval = null, int? maxTopSlowest = 5, int? maxTopHeaviest = 5)
+        /// <param name="maxCacheSizeBytes">Optional maximum cache size in bytes. When specified, enables capacity-based eviction.</param>
+        public BlitzCache(long? defaultMilliseconds = 60000, TimeSpan? cleanupInterval = null, int? maxTopSlowest = 5, int? maxTopHeaviest = 5, long? maxCacheSizeBytes = null)
         {
             if (defaultMilliseconds < 1) throw new ArgumentOutOfRangeException(nameof(defaultMilliseconds), "Default milliseconds must be non-negative");
 
@@ -26,7 +27,7 @@ namespace BlitzCacheCore
             {
                 lock (globalLock)
                 {
-                    globalInstance ??= new BlitzCacheInstance(defaultMilliseconds, cleanupInterval, maxTopSlowest, null, maxTopHeaviest);
+                    globalInstance ??= new BlitzCacheInstance(defaultMilliseconds, cleanupInterval, maxTopSlowest, null, maxTopHeaviest, maxCacheSizeBytes);
                 }
             }
         }